// src/services/openapi/agreement-service/api.ts
export interface RequestByIdDtoLong {
    /**
     * 
     * @type {number}
     * @memberof RequestByIdDtoLong
     */
    id?: number;
}
// ...
export interface RisalahRapatAssignConsentSheetUserRequestDto {
    /**
     * 
     * @type {number}
     * @memberof RisalahRapatAssignConsentSheetUserRequestDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof RisalahRapatAssignConsentSheetUserRequestDto
     */
    assignedTo?: string;
}
// ...
export interface RisalahRapatConsentSheetUserResponseDto {
    /**
     * 
     * @type {number}
     * @memberof RisalahRapatConsentSheetUserResponseDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof RisalahRapatConsentSheetUserResponseDto
     */
    directorateId?: string;
    /**
     * 
     * @type {string}
     * @memberof RisalahRapatConsentSheetUserResponseDto
     */
    directorateLabel?: string;
    /**
     * 
     * @type {number}
     * @memberof RisalahRapatConsentSheetUserResponseDto
     */
    staffId?: number;
    /**
     * 
     * @type {string}
     * @memberof RisalahRapatConsentSheetUserResponseDto
     */
    staffName?: string;
    /**
     * 
     * @type {string}
     * @memberof RisalahRapatConsentSheetUserResponseDto
     */
    divisionId?: string;
    /**
     * 
     * @type {string}
     * @memberof RisalahRapatConsentSheetUserResponseDto
     */
    divisionLabel?: string;
    /**
     * 
     * @type {string}
     * @memberof RisalahRapatConsentSheetUserResponseDto
     */
    jobPositionLabel?: string;
    /**
     * 
     * @type {string}
     * @memberof RisalahRapatConsentSheetUserResponseDto
     */
    consentRole?: string;
    /**
     * 
     * @type {string}
     * @memberof RisalahRapatConsentSheetUserResponseDto
     */
    consentRoleLabel?: string;
    /**
     * 
     * @type {SkuResponseDto}
     * @memberof RisalahRapatConsentSheetUserResponseDto
     */
    sku?: SkuResponseDto;
}
/**
 * 
 * @export
 * @interface RisalahRapatDnDConsentSheetRequestDto
 */
export interface RisalahRapatDnDConsentSheetRequestDto {
    /**
     * 
     * @type {number}
     * @memberof RisalahRapatDnDConsentSheetRequestDto
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof RisalahRapatDnDConsentSheetRequestDto
     */
    previousSequence?: number;
    /**
     * 
     * @type {number}
     * @memberof RisalahRapatDnDConsentSheetRequestDto
     */
    nextSequence?: number;
}
// ...
export interface RisalahRapatListConsentSheetByAssignedToRequestDto {
    /**
     * 
     * @type {string}
     * @memberof RisalahRapatListConsentSheetByAssignedToRequestDto
     */
    bucketProcessId?: string;
    /**
     * 
     * @type {string}
     * @memberof RisalahRapatListConsentSheetByAssignedToRequestDto
     */
    module?: string;
    /**
     * 
     * @type {string}
     * @memberof RisalahRapatListConsentSheetByAssignedToRequestDto
     */
    process?: string;
    /**
     * 
     * @type {string}
     * @memberof RisalahRapatListConsentSheetByAssignedToRequestDto
     */
    assignedTo?: string;
}
// ...
export interface RisalahRapatSaveConsentSheetUserRequestDto {
    /**
     * 
     * @type {number}
     * @memberof RisalahRapatSaveConsentSheetUserRequestDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof RisalahRapatSaveConsentSheetUserRequestDto
     */
    bucketProcessId?: string;
    /**
     * 
     * @type {string}
     * @memberof RisalahRapatSaveConsentSheetUserRequestDto
     */
    module?: string;
    /**
     * 
     * @type {string}
     * @memberof RisalahRapatSaveConsentSheetUserRequestDto
     */
    process?: string;
    /**
     * 
     * @type {string}
     * @memberof RisalahRapatSaveConsentSheetUserRequestDto
     */
    directorateId?: string;
    /**
     * 
     * @type {string}
     * @memberof RisalahRapatSaveConsentSheetUserRequestDto
     */
    divisionId?: string;
    /**
     * 
     * @type {number}
     * @memberof RisalahRapatSaveConsentSheetUserRequestDto
     */
    staffId?: number;
    /**
     * 
     * @type {string}
     * @memberof RisalahRapatSaveConsentSheetUserRequestDto
     */
    assignedTo?: string;
    /**
     * 
     * @type {string}
     * @memberof RisalahRapatSaveConsentSheetUserRequestDto
     */
    consentRole?: string;
    /**
     * 
     * @type {SkuRequestDto}
     * @memberof RisalahRapatSaveConsentSheetUserRequestDto
     */
    sku?: SkuRequestDto;
}
// ...
export interface SkuRequestDto {
    /**
     * 
     * @type {string}
     * @memberof SkuRequestDto
     */
    directorateId?: string;
    /**
     * 
     * @type {string}
     * @memberof SkuRequestDto
     */
    divisionId?: string;
    /**
     * 
     * @type {number}
     * @memberof SkuRequestDto
     */
    staffId?: number;
    /**
     * 
     * @type {string}
     * @memberof SkuRequestDto
     */
    skuNo?: string;
    /**
     * 
     * @type {string}
     * @memberof SkuRequestDto
     */
    skuDate?: string;
}
/**
 * 
 * @export
 * @interface SkuResponseDto
 */
export interface SkuResponseDto {
    /**
     * 
     * @type {number}
     * @memberof SkuResponseDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof SkuResponseDto
     */
    directorateId?: string;
    /**
     * 
     * @type {string}
     * @memberof SkuResponseDto
     */
    directorateLabel?: string;
    /**
     * 
     * @type {string}
     * @memberof SkuResponseDto
     */
    divisionId?: string;
    /**
     * 
     * @type {string}
     * @memberof SkuResponseDto
     */
    divisionLabel?: string;
    /**
     * 
     * @type {string}
     * @memberof SkuResponseDto
     */
    jobPositionLabel?: string;
    /**
     * 
     * @type {number}
     * @memberof SkuResponseDto
     */
    staffId?: number;
    /**
     * 
     * @type {string}
     * @memberof SkuResponseDto
     */
    staffName?: string;
    /**
     * 
     * @type {string}
     * @memberof SkuResponseDto
     */
    skuNo?: string;
    /**
     * 
     * @type {string}
     * @memberof SkuResponseDto
     */
    skuDate?: string;
}
// ...
export class RisalahRapatConsentSheetControllerApi extends BaseAPI {
    /**
     * 
     * @summary API to Assign Consent Sheet User
     * @param {RisalahRapatAssignConsentSheetUserRequestDto} risalahRapatAssignConsentSheetUserRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof RisalahRapatConsentSheetControllerApi
     */
    public assignConsentSheetUser(risalahRapatAssignConsentSheetUserRequestDto: RisalahRapatAssignConsentSheetUserRequestDto, options?: any) {
        return RisalahRapatConsentSheetControllerApiFp(this.configuration).assignConsentSheetUser(risalahRapatAssignConsentSheetUserRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to Get Detail of Consent Sheet User
     * @param {RequestByIdDtoLong} requestByIdDtoLong 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof RisalahRapatConsentSheetControllerApi
     */
    public detailConsentSheetUser(requestByIdDtoLong: RequestByIdDtoLong, options?: any) {
        return RisalahRapatConsentSheetControllerApiFp(this.configuration).detailConsentSheetUser(requestByIdDtoLong, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to Drag n Drop Consent Sheet User List
     * @param {RisalahRapatDnDConsentSheetRequestDto} risalahRapatDnDConsentSheetRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof RisalahRapatConsentSheetControllerApi
     */
    public dndConsentSheetUser(risalahRapatDnDConsentSheetRequestDto: RisalahRapatDnDConsentSheetRequestDto, options?: any) {
        return RisalahRapatConsentSheetControllerApiFp(this.configuration).dndConsentSheetUser(risalahRapatDnDConsentSheetRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to Get List Of UC by Assigned To
     * @param {RisalahRapatListConsentSheetByAssignedToRequestDto} risalahRapatListConsentSheetByAssignedToRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof RisalahRapatConsentSheetControllerApi
     */
    public listConsentSheetUserByAssignedTo(risalahRapatListConsentSheetByAssignedToRequestDto: RisalahRapatListConsentSheetByAssignedToRequestDto, options?: any) {
        return RisalahRapatConsentSheetControllerApiFp(this.configuration).listConsentSheetUserByAssignedTo(risalahRapatListConsentSheetByAssignedToRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to Save Consent Sheet User
     * @param {RisalahRapatSaveConsentSheetUserRequestDto} risalahRapatSaveConsentSheetUserRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof RisalahRapatConsentSheetControllerApi
     */
    public saveConsentSheetUser(risalahRapatSaveConsentSheetUserRequestDto: RisalahRapatSaveConsentSheetUserRequestDto, options?: any) {
        return RisalahRapatConsentSheetControllerApiFp(this.configuration).saveConsentSheetUser(risalahRapatSaveConsentSheetUserRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

}